+++ android/app/src/main/kotlin/com/liso/android/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.liso.app
+package com.liso.android
 
 import io.flutter.embedding.android.FlutterFragmentActivity
 import io.flutter.embedding.engine.FlutterEngine