+++ android/app/src/main/kotlin/com/liso/liso/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.liso.liso
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}