+++ lib/core/translations/en.dart
@@ -118,6 +118,7 @@ final en = {
   "otp_generator": "OTP Generator",
   "seed_generator": "Seed Generator",
   "password_generator": "Password Generator",
+  "weak_passwords": "Weak Passwords",
   "all": "All",
   "okay": "Okay",
   "community_help": "Community & Help",