+++ lib/features/items/items.controller.dart
@@ -62,9 +62,9 @@ class ItemsController extends GetxController with ConsoleMixin, StateMixin {
     } else if (drawerController.filterProtected.value) {
       filteredItems =
           filteredItems.where((e) => e.protected && !e.deleted && !e.trashed);
-    } else if (drawerController.filterPasswordHealth.value) {
+    } else if (drawerController.filterWeakPasswords.value) {
       filteredItems = filteredItems
-          .where((e) => e.hasFragilePasswords && !e.deleted && !e.trashed);
+          .where((e) => e.hasWeakPasswords && !e.deleted && !e.trashed);
     } else if (drawerController.filterTrashed.value) {
       filteredItems = filteredItems.where((e) => e.trashed && !e.deleted);
     } else if (drawerController.filterDeleted.value) {